[#313] AppDelegate 파일 정리
